major refactor. Everything is probably broken, will fix next
