+++ peepingtom/__init__.py
@@ -1,4 +1,4 @@
 from ._version import __version__
 from .io_ import read, write
 from .peeper import Peeper, peep
-from . import analysis, core
+from .functions import merge, stack
