+++ peepingtom/alchemists/__init__.py
@@ -1,2 +1,2 @@
-from .base import Alchemist
+from .alchemist import Alchemist
 from .from_points import PointToLineAlchemist, PointToParticleAlchemist
