+++ peepingtom/datablocks/multiblocks/__init__.py
@@ -1,0 +1,5 @@
+from .multiblock import MultiBlock
+from .dipoleblock import DipoleBlock
+from .orientedpointblock import OrientedPointBlock
+from .particleblock import ParticleBlock
+from .transformblock import TransformBlock
