+++ peepingtom/datablocks/multiblocks/dipoleblock.py
@@ -4,6 +4,7 @@ from .base import MultiBlock
 from .pointblock import PointBlock
 from .orientationblock import OrientationBlock
 from ...utils import align_vectors
+#from ...peeper import DipoleDepictor
 
 
 class DipoleBlock(MultiBlock):
@@ -11,6 +12,7 @@ class DipoleBlock(MultiBlock):
     A DipoleBlock represents a set of n dipoles in an m-dimensional space
     defined by start points and end points
     """
+#    _depictor_class = DipoleDepictor
 
     def __init__(self, startpoints: np.ndarray, endpoints: np.ndarray, **kwargs):
         super().__init__(**kwargs)
