+++ peepingtom/datablocks/simpleblocks/__init__.py
@@ -1,11 +1,7 @@
-from .base import DataBlock, SimpleBlock, MultiBlock
-from .dipoleblock import DipoleBlock
 from .imageblock import ImageBlock
 from .lineblock import LineBlock
 from .meshblock import MeshBlock
 from .orientationblock import OrientationBlock
-from .particleblock import ParticleBlock
 from .pointblock import PointBlock
 from .propertyblock import PropertyBlock
 from .sphereblock import SphereBlock
-from .transformblock import TransformBlock
