+++ peepingtom/depictors/_tests/test_images.py
