+++ peepingtom/depictors/_tests/test_lines.py
