+++ peepingtom/depictors/_tests/test_particles.py
