+++ peepingtom/depictors/_tests/test_points.py
