+++ peepingtom/depictors/dipoles.py
@@ -1,0 +1,4 @@
+from .base import Depictor
+
+class DipoleDepictor(Depictor):
+    pass
