+++ peepingtom/depictors/images.py
