+++ peepingtom/depictors/lines.py
