+++ peepingtom/depictors/particles.py
