+++ peepingtom/depictors/peeper/base.py
