+++ peepingtom/depictors/points.py
