+++ peepingtom/gui/particle_selector.py
