--- peepingtom/peeper/__init__.py
@@ -1,2 +1,0 @@
-from .depictors import Depictor, ImageDepictor, ParticleDepictor, LineDepictor, PointDepictor
-from .peeper import Peeper, peep
