+++ setup.cfg
@@ -50,7 +50,7 @@ install_requires =
 
 [options.entry_points]
 console_scripts =
-    peep = peepingtom.entry_points.peep:cli
+    peep = peepingtom.__main__:cli
 
 
 [options.package_data]
